Add aliases, change file structure, delete useless components
