+++ babel.config.js
@@ -18,8 +18,13 @@ module.exports = {
           '@screens': './src/screens',
           '@services': './src/services',
           '@utils': './src/utils',
-          // Usefull aliases:
+          // Usefull aliases:+
+          '@buttons': './src/components/buttons',
           '@commonComponents': './src/components/common',
+          '@formatters': './src/components/formatters',
+          '@icons': './src/components/icons',
+          '@layout': './src/components/layout/index',
+          '@tokens': './src/components/tokens',
         },
       },
     ],
