+++ src/components/buttons/Button/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ActivityIndicator, StyleSheet, Text, View } from 'react-native';
 import animationScales from '../../../utils/animationScales';
-import Touchable from '../../animations/Touchable';
+import Touchable from '@commonComponents/Touchable';
 import Icon from '../../icons';
 import styles from './styles';
