+++ src/components/common/AccountInfo/index.js
@@ -5,7 +5,7 @@ import { useSelector } from 'react-redux';
 
 import shortAddress from '@utils/shortAddress';
 import { FontStyles } from '@constants/theme';
-import Touchable from '@components/animations/Touchable';
+import Touchable from '@commonComponents/Touchable';
 
 import CopiedToast from '../CopiedToast';
 import styles from './styles';
