+++ src/components/common/EmojiSelector/components/EmojiBar/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 
-import Touchable from '../../../../animations/Touchable';
+import Touchable from '@commonComponents/Touchable';
 import { categoryKeys, Categories } from '../../utils';
 import styles from './styles';
