+++ src/components/common/PasswordInput/index.js
@@ -4,7 +4,7 @@ import LinearGradient from 'react-native-linear-gradient';
 
 import animationScales from '../../../utils/animationScales';
 import { Rainbow } from '../../../constants/theme';
-import Touchable from '../../animations/Touchable';
+import Touchable from '@commonComponents/Touchable';
 import Icon from '../../icons/index';
 import TextInput from '../TextInput';
 import styles from './styles';
