+++ src/components/common/UserIcon/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Text, View } from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
 
-import Touchable from '../../animations/Touchable';
+import Touchable from '@commonComponents/Touchable';
 import { Rainbow } from '../../../constants/theme';
 import styles from './styles';
