+++ src/components/tokens/TokenIcon/styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  genericToken: {
+    width: 41,
+    height: 41,
+    textAlign: 'center',
+    borderRadius: 41,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 10,
+    fontWeight: 'bold',
+  },
+  blackBackground: {
+    backgroundColor: 'black',
+  },
+});
