+++ src/components/tokens/TokenItem/styles.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+import { FontStyles } from '@constants/theme';
+
+export default StyleSheet.create({
+  root: {
+    flexDirection: 'row',
+    justifyContent: 'flex-start',
+    width: '100%',
+  },
+  leftContainer: {
+    flexDirection: 'column',
+    justifyContent: 'space-evenly',
+    marginLeft: 10,
+  },
+  value: {
+    ...FontStyles.Normal,
+    marginLeft: 'auto',
+    alignSelf: 'flex-start',
+  },
+});
