+++ src/screens/flows/Send/components/SaveContact/index.js
@@ -8,7 +8,7 @@ import TextInput from '@commonComponents/TextInput';
 import Header from '@commonComponents/Header';
 import { FontStyles } from '@constants/theme';
 import useContacts from '@hooks/useContacts';
-import Modal from '@components/modal';
+import Modal from '@commonComponents/Modal';
 
 import styles from './styles';
