+++ src/screens/tabs/NFTs/index.js
@@ -8,9 +8,9 @@ import ErrorState from '@commonComponents/ErrorState';
 import EmptyState from '@commonComponents/EmptyState';
 import NftItem from '@screens/tabs/components/NftItem';
 import Container from '@commonComponents/Container';
-import Divider from '@commonComponents/Divider';
 import { getNFTs } from '@redux/slices/user';
 import { Colors } from '@constants/theme';
+import { Separator } from '@layout';
 
 import NftDetail from './screens/NftDetail';
 import styles from './styles';
@@ -57,7 +57,7 @@ const NFTs = () => {
       <Container>
         <WalletHeader />
         <Text style={styles.title}>Collectibles</Text>
-        <Divider />
+        <Separator />
         {!collectionsError ? (
           <FlatList
             bounces
