+++ src/screens/tabs/Profile/screens/Settings/components/BiometricUnlock/index.js
@@ -6,11 +6,11 @@ import RainbowButton from '@components/buttons/RainbowButton';
 import PasswordInput from '@commonComponents/PasswordInput';
 import { validatePassword } from '@redux/slices/keyring';
 import { isValidPassword } from '@constants/general';
-import Column from '@components/layout/Column';
+import { Column } from '@layout';
 import Header from '@commonComponents/Header';
 import { FontStyles } from '@constants/theme';
 import useKeychain from '@hooks/useKeychain';
-import Modal from '@components/modal';
+import Modal from '@commonComponents/Modal';
 
 import styles from './styles';
